move to utils, refactor to avoid circular import
